bookmark

+++ putbackHandler.cgi
@@ -16,12 +16,18 @@ def putbackHandle(form, environ):
 
     topic_id = form.getvalue("topic_id", None)
     docno = form.getvalue("docid", None)
+    title = form.getvalue("title",None)
     atn_db  = DBHandler(db_path.atn)
-    atn_db.cur.execute("SELECT state FROM filter_list WHERE topic_id=? AND docno=?", [int(topic_id), docno])
-    ostate, = atn_db.cur.fetchone()
-    if ostate == 2: origin = 'irrelevant'
-    else: origin = 'duplicate'
-    atn_db.cur.execute("DELETE FROM filter_list WHERE topic_id=? AND docno=?", [int(topic_id), docno])
+    if title = "Bookmarked":
+        atn_db.cur.execute("SELECT state FROM filter_list WHERE topic_id=? AND docno=?", [int(topic_id), docno])
+        ostate, = atn_db.cur.fetchone()
+        if ostate == 2: origin = 'irrelevant'
+        else: origin = 'duplicate'
+        atn_db.cur.execute("DELETE FROM filter_list WHERE topic_id=? AND docno=?", [int(topic_id), docno])
+    else:
+        atn_db.cur.execute("UPDATE bookmark SET state=? WHERE topic_id=? AND docno=?", [1,int(topic_id),docno])
+        origin = 'bookmark'
+
     atn_db.commit()
 
     try: mylog.log_putback(username, origin, topic_id, docno)
